Exp Fee Head & Create Staff Validations

+++ Backend/src/main/java/com/project/repository/CreateStaffRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CreateStaffRepository extends JpaRepository<CreateStaffEntity, Integer> {
+public interface CreateStaffRepository extends JpaRepository<CreateStaffEntity, Long> {
+    // You can add custom queries here if needed
 }
-
